+++ .gitignore
@@ -13,7 +13,7 @@ develop-eggs/
 dist/
 downloads/
 eggs/
-migrations/
+migrations
 .eggs/
 lib/
 lib64/
@@ -128,3 +128,6 @@ dmypy.json
 
 # Pyre type checker
 .pyre/
+
+/.migrations
+/.media
