+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+    "python.linting.enabled": true,
+    "python.linting.flake8Enabled": true
+}
