+++ backend/npo_project/admin/FAQ_admin.py
@@ -1,0 +1,8 @@
+from django.contrib import admin
+from ..models import FAQ
+
+
+@admin.register(FAQ)
+class FAQAdmin(admin.ModelAdmin):
+    list_display = ('question', 'answer')
+    search_fields = ('question',)
