+++ backend/npo_project/admin/__init__.py
@@ -1,0 +1,2 @@
+from .FAQ_admin import FAQAdmin
+from .Location_admin import LocationAdmin
