+++ backend/npo_project/models/__init__.py
@@ -1,0 +1,2 @@
+from .FAQ import FAQ
+from .Location import Location
